+++ src/Utils.cs
@@ -282,6 +282,7 @@ namespace DCFApixels.SpriteEditor
         private static Material hdrMaterial;
         private static Material normalMapMaterial;
         private static Material gaussianBlurMaterial;
+        private static Material motionBlurMaterial;
         private static Material noiseMaterial;
         private static Material effectCacheMaterial;
 
@@ -295,6 +296,7 @@ namespace DCFApixels.SpriteEditor
         public static Material Hdr => GetOrCreate(ref hdrMaterial, "Hidden/TextureCompositor/Hdr");
         public static Material NormalMap => GetOrCreate(ref normalMapMaterial, "Hidden/TextureCompositor/NormalMap");
         public static Material GaussianBlur => GetOrCreate(ref gaussianBlurMaterial, "Hidden/TextureCompositor/GaussianBlur");
+        public static Material MotionBlur => GetOrCreate(ref motionBlurMaterial, "Hidden/TextureCompositor/MotionBlur");
         public static Material Noise => GetOrCreate(ref noiseMaterial, "Hidden/TextureCompositor/Noise");
         public static Material EffectCache => GetOrCreate(ref effectCacheMaterial, "Hidden/TextureCompositor/EffectCache");
         public static Material Transform => GetOrCreate(ref transformMaterial, "Hidden/TextureCompositor/Transform");
@@ -329,6 +331,8 @@ namespace DCFApixels.SpriteEditor
             noiseMaterial = null;
             if (gaussianBlurMaterial != null) UnityEngine.Object.DestroyImmediate(gaussianBlurMaterial);
             gaussianBlurMaterial = null;
+            if (motionBlurMaterial != null) UnityEngine.Object.DestroyImmediate(motionBlurMaterial);
+            motionBlurMaterial = null;
             if (effectCacheMaterial != null) UnityEngine.Object.DestroyImmediate(effectCacheMaterial);
             effectCacheMaterial = null;
             if (normalMapMaterial != null) UnityEngine.Object.DestroyImmediate(normalMapMaterial);